+++ packages/sdk-core/src/ConstantFlowAgreementV1.ts
@@ -71,9 +71,9 @@ export default class ConstantFlowAgreementV1 {
             return this._sanitizeFlowInfo(flowData);
         } catch (err) {
             throw new SFError({
-                type: "CFAV1_READ",
-                customMessage: "There was an error getting the flow",
-                errorObject: err,
+                code: "CFAV1_READ",
+                message: "There was an error getting the flow",
+                cause: err,
             });
         }
     };
@@ -97,10 +97,10 @@ export default class ConstantFlowAgreementV1 {
             return this._sanitizeFlowInfo(flowData);
         } catch (err) {
             throw new SFError({
-                type: "CFAV1_READ",
-                customMessage:
+                code: "CFAV1_READ",
+                message:
                     "There was an error getting the account flow information",
-                errorObject: err,
+                cause: err,
             });
         }
     };
@@ -123,9 +123,9 @@ export default class ConstantFlowAgreementV1 {
             ).toString();
         } catch (err) {
             throw new SFError({
-                type: "CFAV1_READ",
-                customMessage: "There was an error getting net flow",
-                errorObject: err,
+                code: "CFAV1_READ",
+                message: "There was an error getting net flow",
+                cause: err,
             });
         }
     };
@@ -155,9 +155,9 @@ export default class ConstantFlowAgreementV1 {
             return this._sanitizeFlowOperatorData(flowOperatorData);
         } catch (err) {
             throw new SFError({
-                type: "CFAV1_READ",
-                customMessage: "There was an error getting flow operator data",
-                errorObject: err,
+                code: "CFAV1_READ",
+                message: "There was an error getting flow operator data",
+                cause: err,
             });
         }
     };
@@ -186,9 +186,9 @@ export default class ConstantFlowAgreementV1 {
             });
         } catch (err) {
             throw new SFError({
-                type: "CFAV1_READ",
-                customMessage: "There was an error getting flow operator data",
-                errorObject: err,
+                code: "CFAV1_READ",
+                message: "There was an error getting flow operator data",
+                cause: err,
             });
         }
     };
@@ -299,15 +299,15 @@ export default class ConstantFlowAgreementV1 {
         const normalizedFlowOperator = normalizeAddress(params.flowOperator);
         if (!isPermissionsClean(params.permissions)) {
             throw new SFError({
-                type: "UNCLEAN_PERMISSIONS",
-                customMessage: "The desired permissions are unclean",
+                code: "UNCLEAN_PERMISSIONS",
+                message: "The desired permissions are unclean",
             });
         }
 
         if (Number(params.flowRateAllowance) < 0) {
             throw new SFError({
-                type: "NEGATIVE_FLOW_ALLOWANCE",
-                customMessage: "No negative flow allowance allowed",
+                code: "NEGATIVE_FLOW_ALLOWANCE",
+                message: "No negative flow allowance allowed",
             });
         }