+++ packages/sdk-core/src/InstantDistributionAgreementV1.ts
@@ -80,9 +80,9 @@ export default class InstantDistributionAgreementV1 {
             };
         } catch (err) {
             throw new SFError({
-                type: "IDAV1_READ",
-                customMessage: "There was an error getting the subscription",
-                errorObject: err,
+                code: "IDAV1_READ",
+                message: "There was an error getting the subscription",
+                cause: err,
             });
         }
     };
@@ -110,9 +110,9 @@ export default class InstantDistributionAgreementV1 {
             };
         } catch (err) {
             throw new SFError({
-                type: "IDAV1_READ",
-                customMessage: "There was an error getting the index",
-                errorObject: err,
+                code: "IDAV1_READ",
+                message: "There was an error getting the index",
+                cause: err,
             });
         }
     };