+++ packages/sdk-core/src/subgraph/SubgraphClient.ts
@@ -30,9 +30,9 @@ export class SubgraphClient {
             );
         } catch (err) {
             throw new SFError({
-                type: "SUBGRAPH_ERROR",
-                customMessage: `Failed call to subgraph with query ${document}`,
-                errorObject: err,
+                code: "SUBGRAPH_ERROR",
+                message: `Failed call to subgraph with query ${document}`,
+                cause: err,
             });
         }
     }