+++ packages/sdk-core/CHANGELOG.md
@@ -5,6 +5,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Changed
+- `SFError` refactor to be more conventional. It inherits `Error` and uses `cause` to wrap internal errors.
+- Remove serialized internal error from the messages of `SFError` (it's now just included in the `.cause` property)
+
+### Breaking
+- `ErrorType` renamed to `ErrorCode`
+- `SFError.type` renamed to `SFError.code`  
+
 ## [0.4.4] - 2022-06-30
 
 ## Added